requirements.txt
lcu_driver
pystray
pillow
flask
flask_cors
psutil